requirements.txt
torch==2.2.2
torchvision==0.17.2
matplotlib==3.8.4
scikit-learn==1.4.2
pillow==10.3.0
numpy==1.26.4
pandas
tqdm

